successfully webpack local weather
