+++ .gitignore
@@ -1,0 +1,11 @@
+# Dependency directories
+/node_modules
+*/node_modules
+
+# lerna files
+/lerna-debug.log
+
+# Lock files
+/package-lock.json
+/yarn.lock
+*desktop.ini
