+++ package.json
@@ -1,0 +1,32 @@
+{
+  "name": "weather-app",
+  "version": "1.0.0",
+  "description": "Weather Application",
+  "main": "src/index.js",
+  "scripts": {
+    "build": "webpack --mode development",
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "watch": "webpack --progress --watch"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/geraldgsh/weather-app.git"
+  },
+  "author": "Gerald Goh",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/geraldgsh/weather-app/issues"
+  },
+  "homepage": "https://github.com/geraldgsh/weather-app#readme",
+  "dependencies": {
+    "lodash": "^4.17.15",
+    "webpack": "^4.41.6",
+    "webpack-cli": "^3.3.11"
+  },
+  "devDependencies": {
+    "eslint": "^6.8.0",
+    "eslint-config-airbnb-base": "^14.0.0",
+    "eslint-plugin-import": "^2.20.1",
+    "url-loader": "^3.0.0"
+  }
+}
