+++ readme.md
@@ -1,1 +1,198 @@
-feature
+$ npm init
+This utility will walk you through creating a package.json file.
+It only covers the most common items, and tries to guess sensible defaults.
+
+See `npm help json` for definitive documentation on these fields
+and exactly what they do.
+
+Use `npm install <pkg>` afterwards to install a package and
+save it as a dependency in the package.json file.
+
+Press ^C at any time to quit.
+package name: (weather-app)
+version: (1.0.0)
+description: Weather Application
+entry point: (webpack.config.js) src/index.js
+test command:
+git repository: (https://github.com/geraldgsh/weather-app.git)
+keywords:
+author: Gerald Goh
+license: (ISC)
+About to write to /mnt/d/Google_Drive/Microverse/5.Javascript/5.API/weather-app/package.json:
+
+{
+  "name": "weather-app",
+  "version": "1.0.0",
+  "description": "Weather Application",
+  "main": "src/index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/geraldgsh/weather-app.git"
+  },
+  "author": "Gerald Goh",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/geraldgsh/weather-app/issues"
+  },
+  "homepage": "https://github.com/geraldgsh/weather-app#readme"
+}
+
+$ npm install eslint
+npm notice created a lockfile as package-lock.json. You should commit this file.
++ eslint@6.8.0
+added 123 packages from 83 contributors and audited 179 packages in 8.979s
+
+3 packages are looking for funding
+  run `npm fund` for details
+
+found 0 vulnerabilities
+
+$ eslint --init
+? How would you like to use ESLint? To check syntax, find problems, and enforce code style
+? What type of modules does your project use? JavaScript modules (import/export)
+? Which framework does your project use? None of these
+? Does your project use TypeScript? No
+? Where does your code run? Browser
+? How would you like to define a style for your project? Use a popular style guide
+? Which style guide do you want to follow? Airbnb: https://github.com/airbnb/javascript
+? What format do you want your config file to be in? JSON
+Checking peerDependencies of eslint-config-airbnb-base@latest
+Local ESLint installation not found.
+The config that you've selected requires the following dependencies:
+
+eslint-config-airbnb-base@latest eslint@^5.16.0 || ^6.1.0 eslint-plugin-import@^2.18.2
+? Would you like to install them now with npm? Yes
+Installing eslint-config-airbnb-base@latest, eslint@^5.16.0 || ^6.1.0, eslint-plugin-import@^2.18.2
+npm notice save eslint is being moved from dependencies to devDependencies
++ eslint-config-airbnb-base@14.0.0
++ eslint@6.8.0
++ eslint-plugin-import@2.20.1
+added 61 packages from 37 contributors, updated 1 package and audited 387 packages in 8.597s
+
+16 packages are looking for funding
+  run `npm fund` for details
+
+found 0 vulnerabilities
+
+Successfully created .eslintrc.json file in 
+/weather-app
+ESLint was installed locally. We recommend using this local copy instead of your globally-installed copy.
+
+Basic Setup
+First let's create a directory, initialize npm, install webpack locally, and install the webpack-cli (the tool used to run webpack on the command line):
+
+mkdir webpack-demo
+
+npm init -y
+$ npm install webpack webpack-cli --save-devnpm WARN optional SKIPPING OPTIONAL DEPENDENCY: fsevents@1.2.11 (node_modules/fsevents):
+npm WARN notsup SKIPPING OPTIONAL DEPENDENCY: Unsupported platform for fsevents@1.2.11: wanted {"os":"darwin","arch":"any"} (current: {"os":"linux","arch":"x64"})
+
++ webpack-cli@3.3.11
++ webpack@4.41.6
+added 360 packages from 209 contributors, removed 1 package and audited 5675 packages in 25.663s
+
+20 packages are looking for funding
+  run `npm fund` for details
+
+found 0 vulnerabilities
+
+Throughout the Guides we will use diff blocks to show you what changes we're making to directories, files, and code.
+
+Now we'll create the following directory structure, files and their contents:
+
+Create;
+./webpack.config.js
+
+const path = require('path');
+
+module.exports = {
+  mode: 'production',
+  entry: './src/index.js',
+  output: {
+    filename: 'main.js',
+    path: path.resolve(__dirname, 'dist'),
+  },
+  module: {
+    rules: [
+      {
+        test: /\.css$/,
+        use: [
+          'style-loader',
+          'css-loader',
+        ],
+      },
+      {
+        test: /\.(png|svg|jpg|gif|jpeg)$/,
+        use: [
+          'file-loader',
+        ],
+      },
+    ],
+  },
+};
+
+$ yarn build
+yarn run v1.21.1
+$ webpack --mode development
+Hash: 87bba94788336242806f
+Version: webpack 4.41.6
+Time: 102ms
+Built at: 02/21/2020 5:33:18 AM
+  Asset      Size  Chunks             Chunk Names
+main.js  12.3 KiB    main  [emitted]  main
+Entrypoint main = main.js
+[./src/index.js] 1.23 KiB {main} [built]
+[./src/modules/DOMControl.js] 3.12 KiB {main} [built]
+[./src/modules/localControl.js] 1.6 KiB {main} [built]
+Done in 2.67s.
+
+$ yarn watch
+
+yarn run v1.21.1
+$ webpack --watch
+
+webpack is watching the files…
+
+Hash: c1487def981e257fe947
+Version: webpack 4.41.6
+Time: 392ms
+Built at: 02/21/2020 5:34:39 AM
+  Asset      Size  Chunks             Chunk Names
+main.js  1.55 KiB       0  [emitted]  main
+Entrypoint main = main.js
+[0] ./src/index.js + 2 modules 5.96 KiB {0} [built]
+    | ./src/index.js 1.23 KiB [built]
+    | ./src/modules/DOMControl.js 3.12 KiB [built]
+    | ./src/modules/localControl.js 1.6 KiB [built]
+
+
+
+
+$ npm install --save-dev json-loader
+npm WARN rollback Rolling back readable-stream@2.3.6 failed (this is probably harmless): EINVAL: invalid argument, scandir '/mnt/d/Google_Drive/Microverse/5.Javascript/5.API/weather-app/node_modules/fsevents/node_modules'
+npm WARN optional SKIPPING OPTIONAL DEPENDENCY: fsevents@1.2.11 (node_modules/fsevents):
+npm WARN notsup SKIPPING OPTIONAL DEPENDENCY: Unsupported platform for fsevents@1.2.11: wanted {"os":"darwin","arch":"any"} (current: {"os":"linux","arch":"x64"})
+
++ json-loader@0.5.7
+added 1 package from 1 contributor, removed 60 packages and audited 5676 packages in 4.364s
+
+20 packages are looking for funding
+  run `npm fund` for details
+
+found 0 vulnerabilities
+
+webpack.config.js
+
+module.exports = {
+  module: {
+    loaders: [
+      {
+        test: /\.json$/,
+        loader: 'json-loader'
+      }
+    ]
+  }
+}
